make camo configurable
