+++ conf/base.conf
@@ -97,6 +97,10 @@ blog {
     email = ${net.email}
   }
 }
+camo {
+  endpoint = "https://camo.lichess.ovh"
+  secret = ""
+}
 qa {
   collection {
     question = qa_question
